add resize observer
