--- addon/components/pix-app-layout.hbs
@@ -1,12 +1,0 @@
-<div class={{this.classNames}} ...attributes>
-  <section
-    class="pix-app-layout__banner"
-    id="pix-layout-banner-container"
-    {{on-window-resize this.handleMarginContainerNavigation}}
-  >
-    {{yield to="banner"}}
-  </section>
-  <div class="pix-app-layout__navigation">{{yield to="navigation"}}</div>
-  <div class="pix-app-layout__main">{{yield to="main"}}</div>
-  <div class="pix-app-layout__footer">{{yield to="footer"}}</div>
-</div>
