+++ app/modifiers/on-escape-action.js
@@ -1,23 +1,1 @@
-import { modifier } from 'ember-modifier';
-
-export default modifier((element, [callback, focusId = null]) => {
-  function handleKeyUp(event) {
-    const ESCAPE_KEY = 'Escape';
-
-    if (event.key !== ESCAPE_KEY) {
-      return;
-    }
-
-    callback(event);
-
-    if (focusId) {
-      document.getElementById(focusId).focus();
-    }
-  }
-
-  element.addEventListener('keyup', handleKeyUp);
-
-  return () => {
-    element.removeEventListener('keyup', handleKeyUp);
-  };
-});
+export { default } from '@1024pix/pix-ui/modifiers/on-escape-action';
