+++ app/modifiers/on-window-resize.js
@@ -1,12 +1,1 @@
-import { modifier } from 'ember-modifier';
-
-export default modifier(function onWindowResize(element, [action]) {
-  const actionWithElement = () => action(element);
-
-  window.addEventListener('resize', actionWithElement);
-  actionWithElement();
-
-  return () => {
-    window.removeEventListener('resize', actionWithElement);
-  };
-});
+export { default } from '@1024pix/pix-ui/modifiers/on-window-resize';
